\name{Add_numbering}
\alias{Add_numbering}
\title{Add Numbering}
\description{
This addin adds numbering to headers based on their level.
}
\usage{
Add_numbering()
}
\details{
Addin for processing the active document's text and adding numbering to headers.
}
\value{
The modified document text with added numbering to headers.
}

