\name{Remove_numbering}
\alias{Remove_numbering}
\title{Remove Numbering}
\description{
This addin removes numbering from headers.
}
\usage{
Remove_numbering()
}
\details{
Addin for processing the active document's text and removing numbering to headers.
}
\value{
The modified document text with numbering removed from headers.
}
